Factor out decision trees to be imported as json files.
